Cast error az_result values back down to int32_t.

+++ sdk/inc/azure/core/az_result.h
@@ -36,7 +36,7 @@ enum
 };
 
 #define _az_RESULT_MAKE_ERROR(facility, code) \
-  ((uint32_t)_az_ERROR_FLAG | ((uint32_t)(facility) << 16U) | (uint32_t)(code))
+  ((int32_t)((uint32_t)_az_ERROR_FLAG | ((uint32_t)(facility) << 16U) | (uint32_t)(code)))
 
 #define _az_RESULT_MAKE_SUCCESS(facility, code) (((uint32_t)(facility) << 16U) | (uint32_t)(code))
